use long id

+++ src/main/java/com/glodblock/github/extendedae/client/gui/widget/AssemblerMatrixSlot.java
@@ -10,10 +10,10 @@ import net.minecraft.world.item.ItemStack;
 public class AssemblerMatrixSlot extends AppEngSlot {
 
     private final AppEngInternalInventory machineInv;
-    private final int id;
+    private final long id;
     private final int offset;
 
-    public AssemblerMatrixSlot(AppEngInternalInventory machineInv, int machineInvSlot, int offset, int id, int x, int y) {
+    public AssemblerMatrixSlot(AppEngInternalInventory machineInv, int machineInvSlot, int offset, long id, int x, int y) {
         super(machineInv, machineInvSlot);
         this.machineInv = machineInv;
         this.id = id;
@@ -26,7 +26,7 @@ public class AssemblerMatrixSlot extends AppEngSlot {
         return this.getSlotIndex() + this.offset;
     }
 
-    public int getID() {
+    public long getID() {
         return this.id;
     }
